Remove `SEDrive.Conn.Sources.uk :: Cache.t`
After thinking more about how the fact that the UK is this messed up has caused so much harm to real people, I feel bad about adding this as a sort-of joke. Hopefully it's clear than I never meant to brush anything off, but I did see the headlines and decide to add it just for fun. 
I'm well aware that retracting it has essentially no impact on anything (it's still in the history), but it wouldn't seem right to me if I just left it in there.